pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zai-limits-cli"
version = "0.1.0"
description = "CLI to inspect z.ai code plan limits"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
  "rich>=13.9.0",
]

[project.scripts]
zai-limits = "zai_limits_cli.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
